requirements.txt
numpy>=1.24.3,<2.0.0
opencv-python-headless>=4.8.0,<5.0.0
torch>=2.0.1,<3.0.0
torchvision>=0.15.2,<1.0.0
tqdm>=4.66.0,<5.0.0
ffmpeg-python>=0.2.0,<1.0.0
pyyaml>=6.0.0,<7.0.0
numba>=0.58.0,<1.0.0
multiprocessing-logging>=0.3.0,<1.0.0
# MiDaS torch hub dependencies
timm>=0.9.0,<1.0.0
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
httpx>=0.25.0,<1.0.0
setuptools>=65.0.0
streamlit>=1.29.0,<2.0.0